+++ src/Control/Carrier/Interpret.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE FlexibleContexts, FlexibleInstances, FunctionalDependencies, GeneralizedNewtypeDeriving, KindSignatures, RankNTypes, ScopedTypeVariables, TypeApplications, TypeOperators, UndecidableInstances #-}
 
+-- | Provides an 'InterpretC' carrier capable of interpreting an arbitrary effect using a passed-in higher order function to interpret that effect. This is suitable for prototyping new effects quickly.
+
 module Control.Carrier.Interpret
 ( runInterpret
 , runInterpretState
@@ -56,6 +58,8 @@ reify a k =
 -- Note that due to the higher-rank type, you have to use either '$' or explicit application when applying this interpreter. That is, you will need to write @runInterpret f (runInterpret g myPrgram)@ or @runInterpret f $ runInterpret g $ myProgram@. If you try and write @runInterpret f . runInterpret g@, you will unfortunately get a rather scary type error!
 --
 --   prop> run (runInterpret (\ op -> case op of { Get k -> k a ; Put _ k -> k }) get) === a
+--
+-- @since 1.0.0.0
 runInterpret
   :: forall eff m a.
      (HFunctor eff, Monad m)
@@ -82,6 +86,8 @@ runInterpret f m =
 -- | Interpret an effect using a higher-order function with some state variable.
 --
 --   prop> run (runInterpretState (\ s op -> case op of { Get k -> runState s (k s) ; Put s' k -> runState s' k }) a get) === a
+--
+-- @since 1.0.0.0
 runInterpretState
   :: (HFunctor eff, Monad m)
   => (forall x . s -> eff (StateC s m) x -> m (s, x))