+++ src/Control/Carrier/NonDet/Maybe.hs
@@ -1,4 +1,9 @@
 {-# LANGUAGE FlexibleInstances, GeneralizedNewtypeDeriving, MultiParamTypeClasses, TypeOperators, UndecidableInstances #-}
+{- | A carrier for 'NonDet' effects providing choice and failure.
+
+This carrier terminates immediately upon finding a successful result.
+-}
+
 module Control.Carrier.NonDet.Maybe
 ( -- * NonDet effects
   module Control.Effect.NonDet
@@ -23,6 +28,8 @@ import Control.Monad.Trans.Maybe
 --   prop> run (runNonDet empty)    === Nothing
 --   prop> run (runNonDet (pure a)) === Just a
 --   prop> run (runNonDet (let f x = pure x <|> f x in f a)) === Just a
+--
+-- @since 1.0.0.0
 runNonDet :: NonDetC m a -> m (Maybe a)
 runNonDet = runMaybeT . runNonDetC
 {-# INLINE runNonDet #-}
@@ -31,6 +38,8 @@ newtype NonDetC m a = NonDetC { runNonDetC :: MaybeT m a }
   deriving (Alternative, Applicative, Functor, Monad, MonadFix, MonadIO, MonadPlus, MonadTrans)
 
 -- | 'NonDetC' passes 'Fail.MonadFail' operations along to the underlying monad @m@, rather than interpreting it as a synonym for 'empty' à la 'MaybeT'.
+--
+-- @since 1.0.0.0
 instance Fail.MonadFail m => Fail.MonadFail (NonDetC m) where
   fail = lift . Fail.fail
   {-# INLINE fail #-}