+++ src/Control/Carrier/State/Strict.hs
@@ -1,4 +1,9 @@
 {-# LANGUAGE DeriveFunctor, ExplicitForAll, FlexibleContexts, FlexibleInstances, MultiParamTypeClasses, TypeOperators, UndecidableInstances #-}
+
+{- | A carrier for the 'State' effect. It evaluates its inner state strictly, which is the correct choice for the majority of use cases.
+
+Note that the parameter order in 'runState', 'evalState', and 'execState' is reversed compared the equivalent functions provided by @transformers@. This is an intentional decision made to enable the composition of effect handlers with '.' without invoking 'flip'.
+-}
 module Control.Carrier.State.Strict
 ( -- * State effect
   module Control.Effect.State
@@ -24,6 +29,8 @@ import Control.Monad.Trans.Class
 -- | Run a 'State' effect starting from the passed value.
 --
 --   prop> run (runState a (pure b)) === (a, b)
+--
+-- @since 1.0.0.0
 runState :: s -> StateC s m a -> m (s, a)
 runState s x = runStateC x s
 {-# INLINE[3] runState #-}
@@ -31,6 +38,8 @@ runState s x = runStateC x s
 -- | Run a 'State' effect, yielding the result value and discarding the final state.
 --
 --   prop> run (evalState a (pure b)) === b
+--
+-- @since 1.0.0.0
 evalState :: forall s m a . Functor m => s -> StateC s m a -> m a
 evalState s = fmap snd . runState s
 {-# INLINE[3] evalState #-}
@@ -38,11 +47,14 @@ evalState s = fmap snd . runState s
 -- | Run a 'State' effect, yielding the final state and discarding the return value.
 --
 --   prop> run (execState a (pure b)) === a
+--
+-- @since 1.0.0.0
 execState :: forall s m a . Functor m => s -> StateC s m a -> m s
 execState s = fmap fst . runState s
 {-# INLINE[3] execState #-}
 
 
+-- | @since 1.0.0.0
 newtype StateC s m a = StateC { runStateC :: s -> m (s, a) }
   deriving (Functor)