+++ src/Control/Effect/Empty.hs
@@ -1,4 +1,15 @@
 {-# LANGUAGE DeriveFunctor, DeriveGeneric, FlexibleContexts, KindSignatures #-}
+
+{- | An effect modelling nondeterminism without choice (success or failure).
+
+This can be seen as similar to 'Control.Effect.Fail.Fail', but without an error message. The 'Control.Effect.NonDet.NonDet' effect is the composition of 'Empty' and 'Control.Effect.Choice.Choice'.
+
+Predefined carriers:
+
+* "Control.Carrier.Empty.Maybe".
+* If 'Empty' is the last effect in a stack, it can be interpreted directly to a 'Maybe'.
+-}
+
 module Control.Effect.Empty
 ( -- * Empty effect
   Empty(..)
@@ -10,12 +21,10 @@ module Control.Effect.Empty
 , Has
 ) where
 
-import {-# SOURCE #-} Control.Carrier
+import Control.Carrier
 import GHC.Generics (Generic1)
 
--- | An effect modelling nondeterminism without choice.
---
---   This can be seen as similar to 'Control.Effect.Fail.Fail', but without an error message.
+-- | @since 1.0.0.0
 data Empty (m :: * -> *) k = Empty
   deriving (Functor, Generic1)
 
@@ -29,10 +38,14 @@ instance Effect   Empty
 -- @
 -- 'empty' '>>=' k = 'empty'
 -- @
+--
+-- @since 1.0.0.0
 empty :: Has Empty sig m => m a
 empty = send Empty
 
 -- | Conditional failure, returning only if the condition is 'True'.
+--
+-- @since 1.0.0.0
 guard :: Has Empty sig m => Bool -> m ()
 guard True  = pure ()
 guard False = empty