+++ src/Control/Effect/Error.hs
@@ -1,66 +1,27 @@
-{-# LANGUAGE DeriveFunctor, ExistentialQuantification, FlexibleContexts, StandaloneDeriving #-}
+{-# LANGUAGE DeriveFunctor, ExistentialQuantification, FlexibleContexts, StandaloneDeriving, TypeOperators #-}
+
+{- | An effect modelling catchable failure with a polymorphic error type, the combination of 'Throw' and 'Catch'.
+
+This effect is similar to the traditional @MonadError@ typeclass, though it allows the presence of multiple @Error@ effects in a given effect stack. It offers precise exception handling, rather than the dynamic exception hierarchy provided by the @exceptions@ package. The 'Control.Effect.Resource' effect or the @fused-effects-exceptions@ package may be more suitable for handling dynamic/impure effect handling.
+
+Predefined carriers:
+
+* "Control.Carrier.Error.Either".
+* If 'Error' @e@ is the last effect in a stack, it can be interpreted directly to an 'Either' @e@.
+-}
+
 module Control.Effect.Error
 ( -- * Error effect
-  Error(..)
-, throwError
-, catchError
-  -- * Properties
-, throwError_annihilation
-, catchError_interception
+  Error
+, module Control.Effect.Throw
+, module Control.Effect.Catch
   -- * Re-exports
 , Has
 ) where
 
-import {-# SOURCE #-} Control.Carrier
+import Control.Carrier
+import Control.Effect.Catch hiding (Has)
+import Control.Effect.Throw hiding (Has)
 
-data Error exc m k
-  = Throw exc
-  | forall b . Catch (m b) (exc -> m b) (b -> m k)
-
-deriving instance Functor m => Functor (Error exc m)
-
-instance HFunctor (Error exc) where
-  hmap _ (Throw exc)   = Throw exc
-  hmap f (Catch m h k) = Catch (f m) (f . h) (f . k)
-
-instance Effect (Error exc) where
-  handle _     _       (Throw exc)   = Throw exc
-  handle state handler (Catch m h k) = Catch (handler (m <$ state)) (handler . (<$ state) . h) (handler . fmap k)
-
--- | Throw an error, escaping the current computation up to the nearest 'catchError' (if any).
---
--- 'throwError' annihilates '>>=':
---
--- @
--- 'throwError' e '>>=' k = 'throwError' e
--- @
-throwError :: Has (Error exc) sig m => exc -> m a
-throwError = send . Throw
-
--- | Run a computation which can throw errors with a handler to run on error.
---
--- 'catchError' intercepts 'throwError':
---
--- @
--- 'throwError' e \``catchError`\` f = f e
--- @
---
--- Once consequence of this law is that errors thrown by the handler will escape up to the nearest enclosing 'catchError' (if any).
---
--- Note that this effect does /not/ handle errors thrown from impure contexts such as IO,
--- nor will it handle exceptions thrown from pure code. If you need to handle IO-based errors,
--- consider if 'Control.Effect.Resource' fits your use case; if not, use 'liftIO' with
--- 'Control.Exception.try' or use 'Control.Exception.Catch' from outside the effect invocation.
-catchError :: Has (Error exc) sig m => m a -> (exc -> m a) -> m a
-catchError m h = send (Catch m h pure)
-
-
--- Properties
-
--- | 'throwError' annihilates '>>='.
-throwError_annihilation :: Has (Error e) sig m => (m b -> m b -> prop) -> e -> (a -> m b) -> prop
-throwError_annihilation (===) e k = (throwError e >>= k) === throwError e
-
--- | 'catchError' intercepts 'throwError'.
-catchError_interception :: Has (Error e) sig m => (m a -> m a -> prop) -> e -> (e -> m a) -> prop
-catchError_interception (===) e f = (throwError e `catchError` f) === f e
+-- | @since 0.1.0.0
+type Error e = Throw e :+: Catch e